pyproject.toml
[project]
name = "face-recognition-app"
version = "0.1.0"
description = "Single-machine face recognition web app"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "bcrypt>=4.1.3",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.2.0",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "opencv-python>=4.10.0",
  "Pillow>=10.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "anyio>=4.4.0"
]
insightface = [
  "insightface>=0.7.3",
  "onnxruntime>=1.18.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools]
packages = ["app", "app.api", "app.services"]
